SALTO-1251 create workspace top-level cmd and move clean to it as sub cmd
